+++ CHANGELOG.md
@@ -1,3 +1,37 @@
+## 1.4.2
+
+- Require Dart 3.0 or later.
+- Added support for custom post-auth html page
+
+## 1.4.1
+
+- Require Dart 2.19 or later.
+- Allow latest `package:http`.
+
+## 1.4.0
+
+- Update `README` to include a warning about Flutter application usage.
+- Require Dart 2.17 or later.
+
+#### `googlapis_auth.dart`
+
+- `authenticatedClient` function: added optional `bool closeUnderlyingClient`
+  parameter.
+
+#### `auth_browser.dart` library
+
+- Added `AuthenticationException` and use it instead of `Exception` or
+  `StateError` in many cases where authentication can fail.
+- Added `requestAccessCredentials`, `requestAuthorizationCode`, `revokeConsent`,
+  and `CodeResponse` to support the new
+  [Google Identity Services](https://developers.google.com/identity/oauth2/web/guides/overview).
+- Deprecated `createImplicitBrowserFlow` function.
+
+#### `auth_io.dart` library
+
+- Added an optional `listenPort` parameter to `clientViaUserConsent`
+  and `obtainAccessCredentialsViaUserConsent`. 
+
 ## 1.3.1
 
 - Include `plugin_name` during browser authorization.